cumulative emissions
