--- app/javascript/components/widgets/climate/cumulative-emissions/settings.js
@@ -1,4 +1,0 @@
-export default {
-  year: 2018,
-  variable: 'cumulative_deforestation'
-};
